+++ manifest.yml
@@ -3,6 +3,7 @@
 command: ./scripts/run_app.sh $PORT
 services:
   - notify-aws
+  - notify-template-preview
 instances: 1
 memory: 512M
 env:
