+++ requirements.txt
@@ -6,10 +6,11 @@
 # app requirements
 Flask==0.12.1
 Flask-WeasyPrint==0.5
+Flask-HTTPAuth==3.2.2
 html5lib==1.0b10
 wand==0.4.4
 
-git+https://github.com/alphagov/notifications-utils.git@13.10.0#egg=notifications-utils==13.10.0
+git+https://github.com/alphagov/notifications-utils.git@15.0.2#egg=notifications-utils==15.0.2
 
 # PaaS requirements
 gunicorn
